training/requirements.txt
# Training Requirements for SQL Codegen SLM
# Module 2.1: Training Environment Setup

# Core ML libraries
torch==2.1.2
transformers==4.36.2
peft==0.7.1
accelerate==0.25.0
bitsandbytes==0.41.3
trl==0.7.10

# Data handling
datasets==2.15.0
tokenizers==0.15.0

# Training utilities
wandb==0.16.1
tensorboard==2.15.1
scipy==1.11.4

# Evaluation
scikit-learn==1.3.2
rouge-score==0.1.2

# Google Cloud (for GCP deployment)
google-cloud-storage==2.10.0
google-auth==2.25.2

# Configuration
pyyaml>=6.0
